Fixing tests. Substituted hosts in README updated
