+++ test-config/wdio.common.conf.js
@@ -1,4 +1,4 @@
-export default reportDir => ({
+module.exports = reportDir => ({
   //
   // ====================
   // Runner Configuration
