+++ test-config/wdio.docker.chrome.conf.js
@@ -1,4 +1,5 @@
-import common from './wdio.common.conf.mjs'
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const common = require('./wdio.common.conf.js')
 
 export const config = {
   ...common('docker'),
